+++ src/main/java/com/github/italia/daf/data/DataProvider.java
@@ -1,0 +1,8 @@
+package com.github.italia.daf.data;
+
+import java.io.IOException;
+import java.util.List;
+
+public interface DataProvider {
+    List<EmbeddableData> getList() throws IOException;
+}
