+++ src/main/java/com/github/italia/daf/data/EmbeddableData.java
@@ -1,0 +1,44 @@
+package com.github.italia.daf.data;
+
+import com.google.gson.annotations.SerializedName;
+
+public class EmbeddableData {
+    @SerializedName("iframe_url")
+    private String iframeUrl;
+    private String origin;
+    private String title;
+    private String identifier;
+
+    public String getIframeUrl() {
+        return iframeUrl;
+    }
+
+    public void setIframeUrl(String iframeUrl) {
+        this.iframeUrl = iframeUrl;
+    }
+
+    public String getOrigin() {
+        return origin;
+    }
+
+    public void setOrigin(String origin) {
+        this.origin = origin;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getIdentifier() {
+        return identifier;
+    }
+
+    public void setIdentifier(String identifier) {
+        this.identifier = identifier;
+    }
+}
+
