+++ src/test/java/com/github/italia/daf/utils/DafApiMock.java
@@ -1,5 +1,6 @@
 package com.github.italia.daf.utils;
 
+import com.github.italia.daf.data.DataProvider;
 import com.github.italia.daf.service.ApiServiceTest;
 import com.google.gson.GsonBuilder;
 import spark.Service;
