--- src/test/java/com/github/italia/daf/utils/DataProvider.java
@@ -1,9 +1,0 @@
-package com.github.italia.daf.utils;
-
-import com.github.italia.daf.dafapi.HTTPClient;
-
-import java.util.List;
-
-public interface DataProvider {
-    List<HTTPClient.EmbeddableData> getList();
-}
