+++ src/test/java/com/github/italia/daf/utils/RandomDataProvider.java
@@ -1,18 +1,19 @@
 package com.github.italia.daf.utils;
 
-import com.github.italia.daf.dafapi.HTTPClient;
+import com.github.italia.daf.data.DataProvider;
+import com.github.italia.daf.data.EmbeddableData;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class RandomDataProvider implements DataProvider {
-    private List<HTTPClient.EmbeddableData> fakeList;
+    private List<EmbeddableData> fakeList;
 
     public RandomDataProvider() {
 
         this.fakeList = new ArrayList<>();
         for (int i = 0; i < 5; i++) {
-            final HTTPClient.EmbeddableData d = new HTTPClient.EmbeddableData();
+            final EmbeddableData d = new EmbeddableData();
             d.setIdentifier("identifier_" + i);
             d.setOrigin(i % 2 == 0 ? "metabase" : "superset");
             d.setTitle("title_" + i);
@@ -23,7 +24,7 @@ public class RandomDataProvider implements DataProvider {
     }
 
     @Override
-    public List<HTTPClient.EmbeddableData> getList() {
+    public List<EmbeddableData> getList() {
         return this.fakeList;
     }
 }
